requirements.txt
gensim==3.4.0
nltk==3.2.5
hdbscan==0.8.26
kaggle==1.5.9
matplotlib==3.2.2
matplotlib-venn==0.11.6
numpy==1.19.5
pandas==1.1.4
scipy==1.5.4
sklearn-pandas==1.8.0
stop-words==2018.7.23
tensorflow==2.4.1
termcolor==1.1.0
tokenizers==0.9.4
torch==1.7.0
tqdm==4.41.1
transformers==4.1.1
umap-learn==0.5.0
networkx==2.5
